+++ src/app/web_portal_device_api.cpp
@@ -122,6 +122,8 @@ void handleGetVersion(AsyncWebServerRequest *request) {
 		response->print(HAS_BLE ? "true" : "false");
 		response->print(",\"has_backlight\":");
 		response->print(HAS_BACKLIGHT ? "true" : "false");
+		response->print(",\"display_blank_on_save\":");
+		response->print(DISPLAY_BLANK_ON_SAVE ? "true" : "false");
 
 		#if HAS_DISPLAY
 				// Display screen information